refactor: extract SectionCache from Section to decouple cache I/O from EPUB logic
