+++ lib/Epub/Epub/SectionCache.cpp
@@ -1,0 +1,354 @@
+#include "SectionCache.h"
+
+#include <HalStorage.h>
+#include <Logging.h>
+#include <Serialization.h>
+
+#include "Page.h"
+
+bool SectionCache::openForWrite() {
+  if (!Storage.openFileForWrite("SCC", filePath, file)) {
+    return false;
+  }
+  pageCount = 0;
+  return true;
+}
+
+void SectionCache::writeHeader(const int fontId, const float lineCompression, const bool extraParagraphSpacing,
+                               const uint8_t paragraphAlignment, const uint16_t viewportWidth,
+                               const uint16_t viewportHeight, const bool hyphenationEnabled, const bool embeddedStyle,
+                               const uint8_t imageRendering, const bool focusReadingEnabled) {
+  if (!file) {
+    LOG_DBG("SCC", "File not open for writing header");
+    return;
+  }
+  static_assert(HEADER_SIZE == sizeof(FILE_VERSION) + sizeof(fontId) + sizeof(lineCompression) +
+                                   sizeof(extraParagraphSpacing) + sizeof(paragraphAlignment) + sizeof(viewportWidth) +
+                                   sizeof(viewportHeight) + sizeof(pageCount) + sizeof(hyphenationEnabled) +
+                                   sizeof(embeddedStyle) + sizeof(imageRendering) + sizeof(focusReadingEnabled) +
+                                   sizeof(uint32_t) + sizeof(uint32_t) + sizeof(uint32_t) + sizeof(uint32_t),
+                "Header size mismatch");
+  serialization::writePod(file, FILE_VERSION);
+  serialization::writePod(file, fontId);
+  serialization::writePod(file, lineCompression);
+  serialization::writePod(file, extraParagraphSpacing);
+  serialization::writePod(file, paragraphAlignment);
+  serialization::writePod(file, viewportWidth);
+  serialization::writePod(file, viewportHeight);
+  serialization::writePod(file, hyphenationEnabled);
+  serialization::writePod(file, embeddedStyle);
+  serialization::writePod(file, imageRendering);
+  serialization::writePod(file, focusReadingEnabled);
+  serialization::writePod(file, pageCount);
+  serialization::writePod(file, static_cast<uint32_t>(0));  // LUT offset placeholder
+  serialization::writePod(file, static_cast<uint32_t>(0));  // anchor map offset placeholder
+  serialization::writePod(file, static_cast<uint32_t>(0));  // paragraph LUT offset placeholder
+  serialization::writePod(file, static_cast<uint32_t>(0));  // li LUT offset placeholder
+}
+
+uint32_t SectionCache::writePage(std::unique_ptr<Page> page) {
+  if (!file) {
+    LOG_ERR("SCC", "File not open for writing page %d", pageCount);
+    return 0;
+  }
+
+  const uint32_t position = file.position();
+  if (!page->serialize(file)) {
+    LOG_ERR("SCC", "Failed to serialize page %d", pageCount);
+    return 0;
+  }
+  LOG_DBG("SCC", "Page %d processed", pageCount);
+
+  pageCount++;
+  return position;
+}
+
+bool SectionCache::finalize(const std::vector<PageLutEntry>& lut,
+                            const std::vector<std::pair<std::string, uint16_t>>& anchors) {
+  if (!file) {
+    LOG_ERR("SCC", "File not open for finalize");
+    return false;
+  }
+
+  const uint32_t lutOffset = file.position();
+  for (const auto& entry : lut) {
+    if (entry.fileOffset == 0) {
+      LOG_ERR("SCC", "Failed to write LUT due to invalid page positions");
+      file.close();
+      Storage.remove(filePath.c_str());
+      return false;
+    }
+    serialization::writePod(file, entry.fileOffset);
+  }
+
+  const uint32_t anchorMapOffset = file.position();
+  serialization::writePod(file, static_cast<uint16_t>(anchors.size()));
+  for (const auto& [anchor, page] : anchors) {
+    serialization::writeString(file, anchor);
+    serialization::writePod(file, page);
+  }
+
+  const uint32_t paragraphLutOffset = file.position();
+  serialization::writePod(file, static_cast<uint16_t>(lut.size()));
+  for (const auto& entry : lut) {
+    serialization::writePod(file, entry.paragraphIndex);
+  }
+
+  const uint32_t liLutFileOffset = static_cast<uint32_t>(file.position());
+  for (const auto& entry : lut) {
+    serialization::writePod(file, entry.listItemIndex);
+  }
+
+  file.seek(HEADER_SIZE - sizeof(uint32_t) * 4 - sizeof(pageCount));
+  serialization::writePod(file, pageCount);
+  serialization::writePod(file, lutOffset);
+  serialization::writePod(file, anchorMapOffset);
+  serialization::writePod(file, paragraphLutOffset);
+  serialization::writePod(file, liLutFileOffset);
+  file.close();
+  return true;
+}
+
+void SectionCache::closeAndRemove() {
+  if (file) {
+    file.close();
+  }
+  Storage.remove(filePath.c_str());
+}
+
+bool SectionCache::loadHeader(const int fontId, const float lineCompression, const bool extraParagraphSpacing,
+                              const uint8_t paragraphAlignment, const uint16_t viewportWidth,
+                              const uint16_t viewportHeight, const bool hyphenationEnabled, const bool embeddedStyle,
+                              const uint8_t imageRendering, const bool focusReadingEnabled) {
+  if (!Storage.openFileForRead("SCC", filePath, file)) {
+    return false;
+  }
+
+  {
+    uint8_t version;
+    serialization::readPod(file, version);
+    if (version != FILE_VERSION) {
+      file.close();
+      LOG_ERR("SCC", "Deserialization failed: Unknown version %u", version);
+      clearCache();
+      return false;
+    }
+
+    int fileFontId;
+    uint16_t fileViewportWidth, fileViewportHeight;
+    float fileLineCompression;
+    bool fileExtraParagraphSpacing;
+    uint8_t fileParagraphAlignment;
+    bool fileHyphenationEnabled;
+    bool fileEmbeddedStyle;
+    uint8_t fileImageRendering;
+    bool fileFocusReadingEnabled;
+    serialization::readPod(file, fileFontId);
+    serialization::readPod(file, fileLineCompression);
+    serialization::readPod(file, fileExtraParagraphSpacing);
+    serialization::readPod(file, fileParagraphAlignment);
+    serialization::readPod(file, fileViewportWidth);
+    serialization::readPod(file, fileViewportHeight);
+    serialization::readPod(file, fileHyphenationEnabled);
+    serialization::readPod(file, fileEmbeddedStyle);
+    serialization::readPod(file, fileImageRendering);
+    serialization::readPod(file, fileFocusReadingEnabled);
+
+    if (fontId != fileFontId || lineCompression != fileLineCompression ||
+        extraParagraphSpacing != fileExtraParagraphSpacing || paragraphAlignment != fileParagraphAlignment ||
+        viewportWidth != fileViewportWidth || viewportHeight != fileViewportHeight ||
+        hyphenationEnabled != fileHyphenationEnabled || embeddedStyle != fileEmbeddedStyle ||
+        imageRendering != fileImageRendering || focusReadingEnabled != fileFocusReadingEnabled) {
+      file.close();
+      LOG_ERR("SCC", "Deserialization failed: Parameters do not match");
+      clearCache();
+      return false;
+    }
+  }
+
+  serialization::readPod(file, pageCount);
+  file.close();
+  LOG_DBG("SCC", "Deserialization succeeded: %d pages", pageCount);
+  return true;
+}
+
+std::unique_ptr<Page> SectionCache::loadPage(const int pageIndex) {
+  if (!Storage.openFileForRead("SCC", filePath, file)) {
+    return nullptr;
+  }
+
+  file.seek(HEADER_SIZE - sizeof(uint32_t) * 4);
+  uint32_t lutOffset;
+  serialization::readPod(file, lutOffset);
+  file.seek(lutOffset + sizeof(uint32_t) * pageIndex);
+  uint32_t pagePos;
+  serialization::readPod(file, pagePos);
+  file.seek(pagePos);
+
+  auto page = Page::deserialize(file);
+  file.close();
+  return page;
+}
+
+bool SectionCache::clearCache() const {
+  if (!Storage.exists(filePath.c_str())) {
+    LOG_DBG("SCC", "Cache does not exist, no action needed");
+    return true;
+  }
+
+  if (!Storage.remove(filePath.c_str())) {
+    LOG_ERR("SCC", "Failed to clear cache");
+    return false;
+  }
+
+  LOG_DBG("SCC", "Cache cleared successfully");
+  return true;
+}
+
+std::optional<uint16_t> SectionCache::getPageForAnchor(const std::string& anchor) const {
+  FsFile f;
+  if (!Storage.openFileForRead("SCC", filePath, f)) {
+    return std::nullopt;
+  }
+
+  const uint32_t fileSize = f.size();
+  f.seek(HEADER_SIZE - sizeof(uint32_t) * 3);
+  uint32_t anchorMapOffset;
+  serialization::readPod(f, anchorMapOffset);
+  if (anchorMapOffset == 0 || anchorMapOffset >= fileSize) {
+    return std::nullopt;
+  }
+
+  f.seek(anchorMapOffset);
+  uint16_t count;
+  serialization::readPod(f, count);
+  for (uint16_t i = 0; i < count; i++) {
+    std::string key;
+    uint16_t page;
+    serialization::readString(f, key);
+    serialization::readPod(f, page);
+    if (key == anchor) {
+      return page;
+    }
+  }
+
+  return std::nullopt;
+}
+
+std::optional<uint16_t> SectionCache::getPageForParagraphIndex(const uint16_t pIndex) const {
+  FsFile f;
+  if (!Storage.openFileForRead("SCC", filePath, f)) {
+    return std::nullopt;
+  }
+
+  const uint32_t fileSize = f.size();
+  f.seek(HEADER_SIZE - sizeof(uint32_t) * 2);
+  uint32_t paragraphLutOffset;
+  serialization::readPod(f, paragraphLutOffset);
+  if (paragraphLutOffset == 0 || paragraphLutOffset >= fileSize) {
+    return std::nullopt;
+  }
+
+  f.seek(paragraphLutOffset);
+  uint16_t count;
+  serialization::readPod(f, count);
+  if (count == 0) {
+    return std::nullopt;
+  }
+
+  const uint32_t lutEnd = paragraphLutOffset + sizeof(uint16_t) + count * sizeof(uint16_t);
+  if (lutEnd > fileSize) {
+    return std::nullopt;
+  }
+
+  uint16_t resultPage = count - 1;
+  for (uint16_t i = 0; i < count; i++) {
+    uint16_t pagePIdx;
+    serialization::readPod(f, pagePIdx);
+    if (pagePIdx >= pIndex) {
+      resultPage = i;
+      break;
+    }
+  }
+
+  return resultPage;
+}
+
+std::optional<uint16_t> SectionCache::getParagraphIndexForPage(const uint16_t page) const {
+  FsFile f;
+  if (!Storage.openFileForRead("SCC", filePath, f)) {
+    return std::nullopt;
+  }
+
+  const uint32_t fileSize = f.size();
+  f.seek(HEADER_SIZE - sizeof(uint32_t) * 2);
+  uint32_t paragraphLutOffset;
+  serialization::readPod(f, paragraphLutOffset);
+  if (paragraphLutOffset == 0 || paragraphLutOffset >= fileSize) {
+    return std::nullopt;
+  }
+
+  f.seek(paragraphLutOffset);
+  uint16_t count;
+  serialization::readPod(f, count);
+  if (count == 0 || page >= count) {
+    return std::nullopt;
+  }
+
+  const uint32_t entryEnd = paragraphLutOffset + sizeof(uint16_t) + (page + 1) * sizeof(uint16_t);
+  if (entryEnd > fileSize) {
+    return std::nullopt;
+  }
+
+  f.seek(paragraphLutOffset + sizeof(uint16_t) + page * sizeof(uint16_t));
+  uint16_t pIdx;
+  serialization::readPod(f, pIdx);
+  return pIdx;
+}
+
+std::optional<uint16_t> SectionCache::getPageForListItemIndex(const uint16_t liIndex) const {
+  FsFile f;
+  if (!Storage.openFileForRead("SCC", filePath, f)) {
+    return std::nullopt;
+  }
+
+  const uint32_t fileSize = f.size();
+  f.seek(HEADER_SIZE - sizeof(uint32_t));
+  uint32_t liLutOffset;
+  serialization::readPod(f, liLutOffset);
+  if (liLutOffset == 0 || liLutOffset >= fileSize) {
+    return std::nullopt;
+  }
+
+  f.seek(HEADER_SIZE - sizeof(uint32_t) * 2);
+  uint32_t paragraphLutOffset;
+  serialization::readPod(f, paragraphLutOffset);
+  if (paragraphLutOffset == 0 || paragraphLutOffset >= fileSize) {
+    return std::nullopt;
+  }
+
+  f.seek(paragraphLutOffset);
+  uint16_t count;
+  serialization::readPod(f, count);
+  if (count == 0) {
+    return std::nullopt;
+  }
+
+  const uint32_t lutEnd = liLutOffset + count * sizeof(uint16_t);
+  if (lutEnd > fileSize) {
+    return std::nullopt;
+  }
+
+  f.seek(liLutOffset);
+  uint16_t resultPage = count - 1;
+  for (uint16_t i = 0; i < count; i++) {
+    uint16_t pageLiIdx;
+    serialization::readPod(f, pageLiIdx);
+    if (pageLiIdx >= liIndex) {
+      resultPage = i;
+      break;
+    }
+  }
+
+  return resultPage;
+}
